Add basic skeleton of application
The application skeleton offers two ways of serving the files. Way one
is from an embedded file system from within the binary, way two from a
folder in the current file system.

By default files from the embedded fs are served, but if one passes the
`-resoucesDir` flag, the provided path is used.