Add basic role authorization concept
There is now a basic role authorization concept in place, which allows
to secure certain resources by a role. To test it, there is now also a
dummy `/admin` page, which simply renders the index page, but requires
the admin role.

If a user has the admin role, he is allowed to view pretty much
everything.